rearranging and cleaning up functions
