+++ expressyeaself/organize_data.py
@@ -2,11 +2,240 @@
 This script contains functions to organize and split up data based
 on several experimental parameters.
 """
+import expressyeaself.build_promoter as build
+from expressyeaself.utilities import get_seq_count as get_seq_count
+from expressyeaself.utilities import get_time_stamp as get_time_stamp
+from expressyeaself.utilities import smart_open as smart_open
+import gzip
 import os
 import pandas as pd
-from expressyeaself.utilities import smart_open as smart_open
-from expressyeaself.utilities import get_time_stamp as get_time_stamp
+import time as t
 import xlrd
+
+def process_raw_data(input_seqs, scaffold_type=None, homogeneous=False,
+                    deflank=True, insert_into_scaffold=True, extra_padding=0,
+                    pad_front=False, report_loss=True, report_times=True,
+                    remove_files=True):
+    """
+    A wrapper function that:
+    Takes raw data as retrieved from Carl de Boer's publication
+    at https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104878,
+    and processes the sequences according to the custom arguments,
+    pads them to same length, and writes them to an output file along
+    with their expression levels (tab separated). The end of the file
+    contains comments specifying the number of sequences in the file
+    and the lengths of the padded sequences.
+
+    Args:
+    -----
+        input_seqs (str) -- the absolute pathname of the file that
+        contains all of the input sequences and their expression
+        levels (tab separated).
+
+        scaffold_type (str) -- the scaffold type (pTpA or Abf1TATA)
+        that the input sequences had their expression levels
+        measured in.
+
+        homogeneous (bool) -- if True, only sequences of modal
+        length will be processed. If False, all sequences will be
+        processed regardless of length. Default: False.
+
+        deflank (bool) -- if True, removes the constant flanking
+        regions of the input sequences. Default: True.
+
+        insert_into_scaffold (bool) -- if True inserts the input
+        sequences into the appropriate scaffold. If False, the
+        sequences are encoded as they are. Default: True.
+
+        extra_padding (int) -- the number of 'P' characters greater
+        than the maximum sequence length to pad each sequence to.
+        Default: 0.
+
+        pad_front (bool) -- whether to pad out the front (left hand
+        side) or end (right hand side) of the sequences. If True,
+        pads the front. Default: False (will pad the end).
+
+        report_loss (bool) -- if True, reports the number of lines of
+        data lost at each step in the process. Default: False.
+
+        report_times (bool) -- if True, reports the time each step in
+        the cleaning process takes. Default: False.
+
+        remove_files (bool) -- if True, will remove the intermediate
+        files created in the process of processing raw data. Default:
+        False (i.e. intermediary files will be kept).
+
+    Returns:
+    -----
+        processed_data (str) -- the absolute path for the file containing
+        processed sequences along with their expression levels.
+    """
+    # Assertions
+    assert isinstance(input_seqs, str), 'Input file path name must be \
+    passed as a string.'
+    assert os.path.exists(input_seqs), 'Input file does not exist.'
+    assert isinstance(scaffold_type, str), 'Scaffold type must \
+    be passed as a string if specified.'
+    assert scaffold_type == 'pTpA' or scaffold_type == 'Abf1TATA', 'Scaffold\
+    type must be specified as either "pTpA" or "Abf1TATA".'
+    assert isinstance(homogeneous, bool), 'The homogeneous argument must be\
+    passed as a bool.'
+    assert isinstance(deflank, bool), 'The deflank argument must be passed\
+    as a bool.'
+    assert isinstance(insert_into_scaffold, bool), 'The insert_into_scaffold\
+    argument must be passed as a bool.'
+    assert isinstance(extra_padding, int), 'The number of extra vectors to\
+    pad each sequence by should be passed as an integer.'
+    assert extra_padding >= 0, 'extra_padding must be passed as a non-\
+    negative integer.'
+    assert isinstance(pad_front, bool), 'The pad_front argument must be\
+    passed as a bool.'
+    assert isinstance(report_loss, bool), 'The report_loss argument must be\
+    passed as a bool.'
+    assert isinstance(report_times, bool), 'The report_times argument must\
+    be passed as a bool.'
+    assert isinstance(remove_files, bool), 'The remove_files argument must\
+    be passed as a bool.'
+    # Functionality
+    print('Starting processing of raw data...')
+    raw_data = input_seqs
+    # Define final output file path
+    time_stamp = get_time_stamp()
+    relative_path = 'example/processed_data/' + time_stamp
+    processed_data = os.path.join(os.getcwd(), relative_path)
+    # Create log file to write reports to
+    if report_loss or report_times:
+        report = smart_open(processed_data + '_process_report' + '.txt', 'w')
+    # Initialize custom operations if specified (i.e loss + timing reports)
+    if report_loss:
+        loss_report = {}
+        loss_report['Raw Data'] = get_seq_count(input_seqs)
+    if report_times:
+        t_init = t.time()
+        t0 = t_init
+    if remove_files:
+        created_files = [] # keep track of the intermediate files created.
+    # Create new file of only homogeneous (same length) seqs
+    if homogeneous:
+        print('Pulling homogeneous sequences from input file...')
+        input_seqs = pull_homogeneous_seqs(input_seqs,
+                                            scaffold_type=scaffold_type)
+        processed_data += '_homogeneous'
+        if report_loss:
+            loss_report['Homogeneous Seqs'] = get_seq_count(input_seqs)
+        if report_times:
+            t1 = t.time()
+            text = '\tFile created in %s s' %(t1 - t0)
+            print(text)
+            report.write('Homogeneous sequences pulled...\n' + text + '\n')
+            t0 = t1
+        if remove_files:
+            created_files.append(input_seqs)
+    # Remove all of the flanking regions from the input sequences
+    if deflank:
+        print('Removing flank regions from sequences...')
+        input_seqs = build.remove_flanks_from_all_seqs(input_seqs,
+                                    scaffold_type=scaffold_type)
+        processed_data += '_deflanked'
+        if report_loss:
+            loss_report['Deflanked Seqs'] = get_seq_count(input_seqs)
+        if report_times:
+            t1 = t.time()
+            text = '\tFile created in %s s' %(t1 - t0)
+            print(text)
+            report.write('Sequences deflanked...\n' + text + '\n')
+            t0 = t1
+        if remove_files:
+            created_files.append(input_seqs)
+    processed_data += '_sequences'
+    # Insert sequences into appropriate scaffold
+    if insert_into_scaffold:
+        print('Inserting sequences into %s scaffold...' %(scaffold_type))
+        input_seqs = build.insert_all_seq_into_one_scaffold(input_seqs,
+                                    scaffold_type=scaffold_type)
+        processed_data += '_inserted_into_%s_scaffold' %(scaffold_type)
+        if report_loss:
+            loss_report['Scaffold-Inserted Seqs'] = get_seq_count(input_seqs)
+        if report_times:
+            t1 = t.time()
+            text = '\tFile created in %s s' %(t1 - t0)
+            print(text)
+            report.write('Seqs inserted into '+scaffold_type+' scaffold...\n')
+            report.write(text + '\n')
+            t0 = t1
+        if remove_files:
+            created_files.append(input_seqs)
+    # Pad sequences
+    print('Padding sequences...')
+    input_seqs = build.pad_sequences(input_seqs, pad_front=pad_front,
+                                    extra_padding=extra_padding)
+    processed_data += '_padded'
+    if extra_padding != 0:
+        processed_data += '_%s_extra' %(extra_padding)
+    if report_loss:
+        loss_report['Padded Seqs'] = get_seq_count(input_seqs)
+    if report_times:
+        t1 = t.time()
+        text = '\tFile created in %s s' %(t1 - t0)
+        print(text)
+        report.write('Padded sequences...\n')
+        report.write(text + '\n')
+        t0 = t1
+    # Remove intermediate files created in the process
+    if remove_files:
+        created_files.append(input_seqs)
+    # Rename the final output file to reflect how data has been cleaned.
+    processed_data += '_with_exp_levels.txt'
+    # Report end of process and print final output file locations.
+    if input_seqs != raw_data: # i.e. if data has been processed in some way
+        os.rename(input_seqs, processed_data)
+        # Report end of process and print absolute path of processed data.
+        text = ('\nRaw data successfully processed.\nLocation: %s\n'
+            %(processed_data))
+        print(text)
+        report.write(text)
+    else: # If no processing was performed.
+        text ='\nNo processing performed.\n'
+        text += 'Change processing specifications and try again.'
+        print(text)
+        report.write(text + '\n')
+        text = 'Raw data remains unchanged.'
+        print(text)
+        report.write(text + '\n')
+        text = 'Location : %s' %(raw_data)
+        print(text)
+        report.write(text + '\n')
+    # Write the number of seqs and length of seqs to the start of file
+    write_num_and_len_of_seqs_to_file(input_seqs)
+    # Report loss
+    if report_loss:
+        report.write('\nLine counts at each step of the process:\n')
+        for category in loss_report.keys():
+            curr_count = loss_report[category]
+            if category == 'Raw Data':
+                report.write('\t%s : %s\n' %(category, curr_count))
+                prev_count = curr_count
+            else:
+                report.write('\t%s : %s (%s lines lost since last step)\n'
+                    %(category,curr_count, (prev_count - curr_count)))
+                prev_count = curr_count
+    # Remove intermediate files
+    if remove_files:
+        print('\nRemoving intermediate files...')
+        remove_files(created_files)
+        print('Files successfully removed.')
+    print('Process complete.')
+    # Report total time taken
+    if report_times:
+        t_final = t.time()
+        text = '\nTotal processing time : %s s' %(t_final - t_init)
+        print(text)
+        report.write(text)
+        print('Please find the process report in the same directory as the\
+        output file for reports of data losses and timings.')
+    report.close()
+
+    return processed_data
 
 def separate_seq_and_el_data(line):
     """
@@ -158,7 +387,7 @@ def pull_homogeneous_seqs(input_seqs, scaffold_type=None):
     if scaffold_type == 'pTpA':
         modal_length = 110
     elif scaffold_type == 'Abf1TATA':
-        modal_length == 115
+        modal_length = 115
     else:
         _, _, modal_length = get_max_min_mode_length_of_seqs(input_seqs)
     # Find seqs in input file w/ modal length and write them to output file
@@ -264,6 +493,118 @@ def remove_files(files):
             pass
 
     return
+
+def write_num_and_len_of_seqs_to_file(input_seqs):
+    """
+    Prepends the number of sequences and the length of the
+    sequences in an input file to the first 2 lines of the
+    file. Assumes sequences have been processed so that all
+    sequences have been padded to the same length. The first
+    2 lines of the input file will be in the following format
+    after writing the info to the file:
+    "
+    number_of_seqs_in_file\t<###>
+    length_of_each_sequence\t<$$$>
+    "
+    where '<###>' is the number of sequences in the file, and
+    '<$$$>'is the length to which every sequence in the file is
+    padded.
+
+    Args:
+    -----
+        input_seqs (str) -- the absolute path of the processed
+        input sequences to extract information from.
+
+    Returns:
+    -----
+        None
+    """
+    # Assertions
+    assert isinstance(input_seqs, str), 'Absolute pathname must be passed\
+    as a string.'
+    assert os.path.exists(input_seqs), 'Input file does not exist.'
+    # Functionality
+    num_seqs = get_seq_count(input_seqs)
+    with smart_open(input_seqs, 'r') as f:
+        line = f.readline()
+        seq, _ = organize.separate_seq_and_el_data(line)
+        len_seqs = len(seq) # assumes all sequences padded to same length
+    with smart_open(input_seqs, "r+") as f:
+        contents = f.read()
+    with smart_open(input_seqs, "w+") as f:
+        line_to_append = 'number_of_seqs_in_file\t' + str(num_seqs) + '\n'
+        line_to_append += 'length_of_each_sequence\t' + str(len_seqs) + '\n'
+        if input_seqs.endswith('.gz'):
+            line_to_append = line_to_append.encode()
+        f.write(line_to_append + contents)
+
+    return
+
+def get_num_and_len_of_seqs_from_file(input_seqs):
+    """
+    Returns the number of sequences and length of sequences in an
+    input file. Assumes sequences have been processed so that all
+    sequences have been padded to the same length, and that the
+    file containing the process sequences have the first 2 lines in
+    the following format:
+    "
+    number_of_seqs_in_file\t<###>
+    length_of_each_sequence\t<$$$>
+    "
+    where '<###>' is the number of sequences in the file, and
+    '<$$$>'is the length to which every sequence in the file is
+    padded.
+
+    Args:
+    -----
+        input_seqs (str) -- the absolute path of the processed
+        input sequences to extract information from.
+
+    Returns:
+    -----
+        num_seqs (int) -- the number of sequences in input_seqs.
+
+        len_seqs (int) -- the length of the all the padded
+        sequences in the input file.
+    """
+    # Assertions
+    assert isinstance(input_seqs, str), 'Absolute pathname must be passed\
+    as a string.'
+    assert os.path.exists(input_seqs), 'Input file does not exist.'
+    # Functionality
+    with smart_open(input_seqs, 'r') as f:
+        # Parse first line of file containing info about num of seqs in file
+        first_line = check_valid_line(f.readline())
+        assert first_line != 'skip_line', 'Invalid first line of file. Must\
+        be of the form: "number_of_seqs_in_file\t<###>" where <###> is the\
+        number of sequences in the file.'
+        token, num_seqs = separate_seq_and_el_data(first_line)
+        try:
+            num_seqs = int(num_seqs)
+        except ValueError:
+            raise AssertionError('Number of sequences on first line must be\
+            an integer.')
+        assert token == 'number_of_seqs_in_file', 'First line of the input\
+        file must be of the form: "number_of_seqs_in_file\t<###>" where\
+        <###> is the number of sequences in the file.'
+
+        # Parse 2nd line of file containing info about length of seqs in file
+        second_line = check_valid_line(f.readline())
+        assert second_line != 'skip_line', 'Invalid second line of file.\
+        Must be of the form: "length_of_each_sequence\t<###>" where <###> is\
+        the length of every sequence in the file.'
+        token, len_seqs = separate_seq_and_el_data(second_line)
+        try:
+            len_seqs = int(len_seqs)
+        except ValueError:
+            raise AssertionError('Sequence length on second line must be an\
+            integer.')
+        assert token == 'length_of_each_sequence', 'Second line of the input\
+        file must be of the form: "length_of_each_sequence\t<###>" where\
+        <###> is the length of every sequence in the file. Assumes\
+        homogeneity and/or padding of sequences.'
+
+    return num_seqs, len_seqs
 
 def split_scaffolds_by_type(infile):
     """
